--- src/components/exercise.vue
@@ -1,27 +1,0 @@
-
-<template>
-	<!--练习展示页面-->
-    <div>
-        
-    </div>
-</template>
-
-<script>
-    export default{
-		data(){
-			return{
-				
-			}
-		},
-		methods:{
-			
-		},mounted(){
-			
-		}
-	}
-</script>
-
-<style>
-</style>
-
-
